requirements.txt
# Web framework
Flask>=3.0.0
gunicorn>=21.0.0
WTForms>=3.1.0
Flask-WTF>=1.2.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Cors>=4.0.0

# Database
psycopg2-binary>=2.9.0

# Visualization and computation
matplotlib>=3.8.0
numpy>=1.26.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
